fix: portability bugs blocking packaging CI (macOS, Windows, Android)
- segmented_primedia_spool.cpp: std::max(committedBytes, start + segSize)
  fails to deduce a common type under libc++ (macOS), where size_t and
  uint64_t are distinct 64-bit unsigned types. Cast explicitly.
- lmdb_cache.hpp: mdb_env_open needs a narrow C string, but
  std::filesystem::path::c_str() returns wchar_t* on Windows. Use
  .string().c_str(), matching the existing pattern in android_bootstrap.cpp.
- CMakeLists.txt (Android): vcpkg's libunibreak port ships a CMake package
  config, not a .pc file, so pkg_check_modules could never find it. Locate
  it via find_package(libunibreak CONFIG REQUIRED) instead, per vcpkg's own
  documented usage for this port.

All three were pre-existing, unrelated to this branch's own changes, and
surfaced as CI failures on brew (macOS, SDL3), zip (Windows, MSYS2
UCRT64), and build (Android) once this branch was pushed.

+++ apps/xudu/core/segmented_primedia_spool.cpp
@@ -138,7 +138,8 @@ bool SegmentedPrimediaSpool::addSealedSegment(
   if (arena.isValid() && (start % ps == 0) && (segSize % ps == 0)) {
     mapped = arena.mapFileFixed(arena.base() + start, fd, 0, segSize, false);
     if (mapped) {
-      committedBytes = std::max(committedBytes, start + segSize);
+      committedBytes =
+          std::max(committedBytes, static_cast<std::size_t>(start + segSize));
     }
   }
